GasusBotDashboard/requirements.txt
Django
environs
djangorestframework
psycopg2